Moved ContextExtensions.kt & SingletonHolder.kt into utils package
